N ando patch 2

+++ examples/ConfigSample/VectorConvert.h
@@ -16,7 +16,7 @@ namespace std
     is >> s;
     sv = coil::split(s ,",");
     v.resize(sv.size());
-    for (int i(0), len(sv.size()); i < len; ++i)
+    for (size_t i(0), len(sv.size()); i < len; ++i)
       {
 	T tv;
 	if (coil::stringTo(tv, sv[i].c_str()))
@@ -26,5 +26,5 @@ namespace std
       }
     return is;
   }
-}
+} // namespace std
